feat: move message creation to component
